backend_pre_migrations/requirements.txt
Django==3.2
psycopg2-binary==2.9.5
python-donenv==0.21.1
djangorestframework-3.14.0
djoser=2.1.0
django-filter=22.1
reportlab=3.6.12
drf-extra-fields=3.4.1
